% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{append_subcopula}
\alias{append_subcopula}
\title{add subcopula to a nac_Node}
\usage{
append_subcopula(nac_Node, new_subcopula)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{new_subcopula}{new subcopula to be added}
}
\description{
add subcopula to a nac_Node
}
