\name{AssocMeasures}
\docType{methods}
\alias{kendallsTau}
\alias{spearmansRho}
\alias{tailIndex}
\alias{calibKendallsTau}
\alias{calibSpearmansRho}
  
\alias{kendallsTau-methods}
\alias{kendallsTau,ANY-method}
\alias{kendallsTau,copula-method}
\alias{kendallsTau,archmCopula-method}
\alias{kendallsTau,amhCopula-method}
\alias{kendallsTau,claytonCopula-method}
\alias{kendallsTau,frankCopula-method}
\alias{kendallsTau,gumbelCopula-method}
\alias{kendallsTau,normalCopula-method}
\alias{kendallsTau,tCopula-method}
\alias{kendallsTau,fgmCopula-method} 

\alias{spearmansRho-methods}
\alias{spearmansRho,ANY-method}
\alias{spearmansRho,copula-method}
\alias{spearmansRho,frankCopula-method}
\alias{spearmansRho,normalCopula-method}
\alias{spearmansRho,tCopula-method}
\alias{spearmansRho,fgmCopula-method} 

\alias{tailIndex-methods}
\alias{tailIndex,ANY-method}
\alias{tailIndex,copula-method}
\alias{tailIndex,evCopula-method}
\alias{tailIndex,claytonCopula-method}
\alias{tailIndex,frankCopula-method}
\alias{tailIndex,gumbelCopula-method}
\alias{tailIndex,normalCopula-method}
\alias{tailIndex,tCopula-method}

\alias{calibKendallsTau-methods}
\alias{calibKendallsTau,ANY-method}
\alias{calibKendallsTau,copula-method}
\alias{calibKendallsTau,ellipCopula-method}
\alias{calibKendallsTau,normalCopula-method}
\alias{calibKendallsTau,tCopula-method}
\alias{calibKendallsTau,claytonCopula-method}

\alias{calibSpearmansRho-methods}
\alias{calibSpearmansRho,ANY-method}
\alias{calibSpearmansRho,copula-method}
\alias{calibSpearmansRho,ellipCopula-method}
\alias{calibSpearmansRho,normalCopula-method}
\alias{calibSpearmansRho,tCopula-method}


\title{Association Measures in Package `copula'}
\description{
 Methods to compute association measures such as Kendall's Tau,
 Spearman's Rho, and tail dependence index for bivaraite copulas.
 Calibration is supplied for Kendall's Tau and Spearman's Rho.
}
\section{Methods}{
\describe{
  \item{copula = "copula"}{ Association measure for a \code{"copula"}
    object.}
}}
\keyword{methods}
\keyword{multivariate}
