% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_theta.R
\name{bound_theta}
\alias{bound_theta}
\title{A helper method to ensure the bound of NAC parameters}
\usage{
bound_theta(nac_Node, theta)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{theta}{a parameter to check}
}
\value{
a valid theta
}
\description{
A helper method to ensure the bound of NAC parameters
}
