% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{count_subcopula}
\alias{count_subcopula}
\title{count number of subcopulas}
\usage{
count_subcopula(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\description{
count number of subcopulas
}
