% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNC.R
\name{dNC}
\alias{dNC}
\title{Calculate density of a nested archimedean copula}
\usage{
dNC(U, naclist, family)
}
\arguments{
\item{U}{matrix of pseudo-observations}

\item{naclist}{list specifying nesting structure}
}
\value{
vector of length nrow(U) with density values
}
\description{
Calculate density of a nested archimedean copula
}
