% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dncopula.R
\name{dncopula}
\alias{dncopula}
\title{Calculate density of a nested archimedean copula}
\usage{
dncopula(nac_Node, U, log = TRUE)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{U}{matrix of pseudo-observations}
}
\value{
vector of length nrow(U) with density values
}
\description{
Calculate density of a nested archimedean copula
}
