% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_par.R
\name{estimate_nac}
\alias{estimate_nac}
\title{A helper method to estimate the parameters of one Archimedean copula}
\usage{
estimate_nac(nac_Node, U)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{U}{pseudo-observations}
}
\value{
nac_Node a nested Archimedean copula with new estimates
}
\description{
A helper method to estimate the parameters of one Archimedean copula
}
