% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_par.R
\name{estimate_par}
\alias{estimate_par}
\title{estimate copula parameters inside out}
\usage{
estimate_par(nac_Node, U)
}
\arguments{
\item{nac_Node}{an nac node}

\item{U}{pseudo-observations}
}
\value{
a list of nac node, estimated param, and result of optim function
}
\description{
estimate copula parameters inside out
}
