\name{generator}
\docType{methods}
\alias{genFun}
\alias{genFun-methods}
\alias{genFun,claytonCopula-method}
\alias{genFun,frankCopula-method}
\alias{genFun,gumbelCopula-method}
\alias{genFun,amhCopula-method}

\alias{genInv}
\alias{genInv-methods}
\alias{genInv,claytonCopula-method}
\alias{genInv,frankCopula-method}
\alias{genInv,gumbelCopula-method}
\alias{genInv,amhCopula-method}

\alias{genFunDer1}
\alias{genFunDer1-methods}
\alias{genFunDer1,claytonCopula-method}
\alias{genFunDer1,frankCopula-method}
\alias{genFunDer1,gumbelCopula-method}
\alias{genFunDer1,amhCopula-method}

\alias{genFunDer2}
\alias{genFunDer2-methods}
\alias{genFunDer2,claytonCopula-method}
\alias{genFunDer2,frankCopula-method}
\alias{genFunDer2,gumbelCopula-method}
\alias{genFunDer2,amhCopula-method}

\alias{Afun}
\alias{Afun-methods}
\alias{Afun,galambosCopula-method}
\alias{Afun,gumbelCopula-method}
\alias{Afun,huslerReissCopula-method}

\title{Generators for Archimedean and EV Copulas}
\description{
 Methods function to evaluate generator function, inverse generator
 function, and derivatives of generator function.
}
\section{Methods}{
\describe{
  \item{copula = "copula"}{ Association measure for a \code{"copula"}
    object.}
}
}
\keyword{methods}
\keyword{multivariate}
