% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cor_theta.R
\name{get_cor_theta}
\alias{get_cor_theta}
\title{Get theta for nested copula via U correlation}
\usage{
get_cor_theta(nac_Node, U)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{U}{pseudo-observation}
}
\value{
theta for nested copula based on U correlation
}
\description{
Get theta for nested copula via U correlation
}
