% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{get_family}
\alias{get_family}
\title{get family of nested Archimedean copula}
\usage{
get_family(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\value{
family
}
\description{
get family of nested Archimedean copula
}
