% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iniCor.R
\name{get_iniCor}
\alias{get_iniCor}
\title{get initial theta for nested copula with U}
\usage{
get_iniCor(nac_Node, U)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{U}{pseudo-observation}
}
\value{
initial theta for nested copula based on U correlation
}
\description{
get initial theta for nested copula with U
}
