% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loglik.R
\name{get_loglik}
\alias{get_loglik}
\title{Log likelihood of nested A. copula parameters}
\usage{
get_loglik(nac_Node, U, par)
}
\arguments{
\item{nac_Node}{object of class}

\item{U}{pseudo-observations}

\item{par}{a vector of copula parameters on unbounded scale (real numbers)}
}
\value{
Log-likelihood
}
\description{
Log likelihood of nested A. copula parameters
}
