% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_par.R
\name{get_num_par}
\alias{get_num_par}
\title{calculate number of parameters}
\usage{
get_num_par(nac_Node)
}
\arguments{
\item{nac_Node}{an NAC node}
}
\value{
number of parameters
}
\description{
calculate number of parameters
}
