% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par}
\alias{get_par}
\title{get all parameters of a nested Archimedean copula}
\usage{
get_par(nac_Node)
}
\arguments{
\item{nac_Node}{an NAC node}
}
\value{
par: a vector of NAC parameters
}
\description{
get all parameters of a nested Archimedean copula
}
