% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{get_subcopula}
\alias{get_subcopula}
\title{get subcopulas for nested Archimedean copula, NULL if no child}
\usage{
get_subcopula(nac_Node, index)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{index}{index of subcopula}
}
\value{
subcopulas
}
\description{
get subcopulas for nested Archimedean copula, NULL if no child
}
