% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNCopula.R
\name{get_theta}
\alias{get_theta}
\title{nested copula with U and inner copulas.. get_theta for list with length of 4}
\usage{
get_theta(U, naclist)
}
\description{
nested copula with U and inner copulas.. get_theta for list with length of 4
}
