% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_methods.R
\name{get_U_indices}
\alias{get_U_indices}
\title{get pseudo-obs for nested Archimedean copula}
\usage{
get_U_indices(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\value{
pseudo-obs
}
\description{
get pseudo-obs for nested Archimedean copula
}
