% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xv.R
\name{get_xv}
\alias{get_xv}
\title{Cross Validation for nested A. copulas}
\usage{
get_xv(x, nac_Node, k = NULL)
}
\arguments{
\item{x}{a data matrix that will be transformed to pseudo-observations}

\item{nac_Node}{an NAC node}

\item{k}{the number of data blocks; if k = NULL, nrow(x) blocks are considered}
}
\value{
A real number equal to the cross-validation criterion
}
\description{
Cross Validation for nested A. copulas
}
