% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{has_subcopula}
\alias{has_subcopula}
\title{boolean function to check if NAC has a subcopula}
\usage{
has_subcopula(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\description{
boolean function to check if NAC has a subcopula
}
