% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{is.nac_Node}
\alias{is.nac_Node}
\title{boolean function to check if it's an NAC node}
\usage{
is.nac_Node(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\value{
true false
}
\description{
boolean function to check if it's an NAC node
}
