% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikNC.R
\name{loglikNC}
\alias{loglikNC}
\title{estimation of nested A. copula parameters}
\usage{
loglikNC(par, U, naclist)
}
\arguments{
\item{par}{vector of copula parameters}

\item{U}{observations}

\item{naclist}{list specifying nesting structure}
}
\value{
result obtained from optimizing provided parameters
}
\description{
estimation of nested A. copula parameters
}
