% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncopula-package.R
\docType{package}
\name{ncopula}
\alias{ncopula}
\alias{ncopula-package}
\title{ncopula}
\description{
Utility functions for generating and plotting forecasts from a variety of
   models used by team Kernel of Truth for the influenza forecasting
   challenge run by the CDC.
}
\author{
Amelia Tran
}
