% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{new_nac_node}
\alias{new_nac_node}
\title{create an NAC node}
\usage{
new_nac_node(family = character(), theta = double(),
  U_indices = double(), subcopula = NULL)
}
\arguments{
\item{family}{a nested Archimedean copula family}

\item{theta}{initial copula theta}

\item{U_indices}{indices of pseudo-observation}

\item{subcopula}{child copulas}
}
\value{
NAC node
}
\description{
create an NAC node
}
