% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimNC.R
\name{optimNC}
\alias{optimNC}
\title{Repeat optim function many many times - for a nested copula that belongs to 1 specific family}
\usage{
optimNC(num_of_times, U, naclist)
}
\arguments{
\item{num_of_times}{number of times to run optim function}

\item{U}{matix of pseudo observation}

\item{naclist}{list specifying nesting structure}

\item{num_of_times}{number of times to call optim function on NC}

\item{naclist}{list specifying nesting structure}

\item{U}{pseudo observations}
}
\value{
A largest result obtained from running optim function many times
}
\description{
Repeat optim function many many times - for a nested copula that belongs to 1 specific family
}
