% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNC.R
\name{pNC}
\alias{pNC}
\title{Calculate cdf of a nested archimedean copula}
\usage{
pNC(U, naclist)
}
\arguments{
\item{U}{matrix of pseudo-observations}

\item{naclist}{list specifying nesting structure}
}
\value{
vector of length nrow(U) with cdf values
}
\description{
Calculate cdf of a nested archimedean copula
}
