% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pncopula.R
\name{pncopula}
\alias{pncopula}
\title{Calculate cdf of a nested archimedean copula}
\usage{
pncopula(nac_Node, U)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{U}{matrix of pseudo-observations}
}
\value{
vector of length nrow(U) with cdf values
}
\description{
Calculate cdf of a nested archimedean copula
}
