% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{set_family}
\alias{set_family}
\title{set family for nested Archimedean copula}
\usage{
set_family(nac_Node, new_family)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{new_family}{new family}
}
\description{
set family for nested Archimedean copula
}
