% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_par.R
\name{set_par}
\alias{set_par}
\title{set nested copula with new parameters}
\usage{
set_par(nac_Node, theta, current_index = 1)
}
\arguments{
\item{nac_Node}{an NAC node}

\item{current_index}{index of copula}

\item{par}{a vector of new parameters}
}
\value{
an nac_Node with new parameters and current index
}
\description{
set nested copula with new parameters
}
