% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{set_subcopula}
\alias{set_subcopula}
\title{set new subcopula for nested Archimedean copula}
\usage{
set_subcopula(nac_Node, index, new_subcopula)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{index}{index of subcopula}

\item{new_subcopula}{new subcopula}
}
\description{
set new subcopula for nested Archimedean copula
}
