% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{set_theta}
\alias{set_theta}
\title{set new initial theta for nested Archimedean copula}
\usage{
set_theta(nac_Node, new_theta)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{new_theta}{new theta}
}
\description{
set new initial theta for nested Archimedean copula
}
