% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theta_from_unbounded.R
\name{set_theta_from_unbounded}
\alias{set_theta_from_unbounded}
\title{transform parameters of NAC so that they are within the bounds}
\usage{
set_theta_from_unbounded(nac_Node, par, current_index = 1)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{par}{vector of copula parameters}

\item{current_index}{index of copula}
}
\value{
a list of updated parameters within the bounds and current index
}
\description{
transform parameters of NAC so that they are within the bounds
}
