% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_theta_unbounded.R
\name{set_theta_unbounded}
\alias{set_theta_unbounded}
\title{transform parameters so that they are unbounded}
\usage{
set_theta_unbounded(nac_Node, par, current_index = 1)
}
\arguments{
\item{nac_Node}{a nested A. copula}

\item{par}{vector of copula parameters}

\item{current_index}{index of copula}
}
\value{
a list of unbounded parameters and current index
}
\description{
transform parameters so that they are unbounded
}
