% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_methods.R
\name{set_U_indices}
\alias{set_U_indices}
\title{set new pseudo-obs for nested Archimedean copula}
\usage{
set_U_indices(nac_Node, new_indices)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}

\item{new_indices}{new indices of U}
}
\description{
set new pseudo-obs for nested Archimedean copula
}
