\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,fitCopula-method}
\alias{summary,fitMvdc-method}
\title{  Methods for Function summary in Package `copula' }
\description{
  Methods for function \code{summary} in package \pkg{copula}.
}
\section{Methods}{
  \describe{
    \item{object = "fitCopula"}{ ~~describe this method here }
    
    \item{object = "fitMvdc"}{ ~~describe this method here }
}}
\examples{

}
\keyword{methods}

