% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_par.R
\name{transform_par}
\alias{transform_par}
\title{Transform parameters so that they are within the bounds}
\usage{
transform_par(par, naclist, current_index = 1)
}
\arguments{
\item{par}{vector of copula parameters}

\item{naclist}{list specifying nesting structure}

\item{current_index}{index of copula}
}
\value{
a list of updated parameters within the bounds and current index
}
\description{
Transform parameters so that they are within the bounds
}
