% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_par.R
\name{transform_par_unbounded}
\alias{transform_par_unbounded}
\title{Transform parameters so that they are unbounded}
\usage{
transform_par_unbounded(par, naclist, current_index = 1)
}
\arguments{
\item{par}{vector of copula parameters}

\item{naclist}{list specifying nesting structure}

\item{current_index}{index of copula}
}
\value{
a list of updated parameters within the bounds and current index
}
\description{
Transform parameters so that they are unbounded
}
