% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_theta_bounded_to_unbounded.R
\name{transform_theta_bounded_to_unbounded}
\alias{transform_theta_bounded_to_unbounded}
\title{Transform vector of parameters that are used for each copula in a NAC
to a vector of unbounded real numbers (as optim might use).}
\usage{
transform_theta_bounded_to_unbounded(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\value{
a vector of unbounded parameters
}
\description{
Transform vector of parameters that are used for each copula in a NAC
to a vector of unbounded real numbers (as optim might use).
}
