% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_theta_bounded_to_unbounded.R
\name{untransform_nac}
\alias{untransform_nac}
\title{A helper method to transform vector of NAC parameters
to a vector of unbounded real numbers (as optim might use).}
\usage{
untransform_nac(nac_Node)
}
\arguments{
\item{nac_Node}{a nested Archimedean copula}
}
\value{
an unbounded theta
}
\description{
A helper method to transform vector of NAC parameters
to a vector of unbounded real numbers (as optim might use).
}
