% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_par.R
\name{update_par}
\alias{update_par}
\title{Update nested copula parameters}
\usage{
update_par(nac_Node, par, current_index = 1)
}
\arguments{
\item{nac_Node}{an NAC node}

\item{par}{a vector of new parameters}

\item{current_index}{index of copula}
}
\value{
an updated nac_Node with new parameters and current index
}
\description{
Update nested copula parameters
}
