% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePars.R
\name{updatePars}
\alias{updatePars}
\title{Update nested copula parameters}
\usage{
updatePars(par, naclist, current_index = 1)
}
\arguments{
\item{par}{vector of copula parameters}

\item{naclist}{list specifying nesting structure}

\item{current_index}{index of copula}
}
\value{
a list of updated naclist with new parameters and current index
}
\description{
Update nested copula parameters
}
