% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvNC.R
\name{xvNCopula}
\alias{xvNCopula}
\title{Cross Validation for nested A. copulas}
\usage{
xvNCopula(x, naclist, k = NULL)
}
\arguments{
\item{x}{a data matrix that will be transformed to pseudo-observations}

\item{naclist}{list specifying nesting structure}

\item{k}{the number of data blocks; if k = NULL, nrow(x) blocks are considered}
}
\value{
A real number equal to the cross-validation criterion
}
\description{
Cross Validation for nested A. copulas
}
