\name{AssocMeasures}
\docType{methods}
\alias{kendallsTau}
\alias{kendallsTau-methods}
\alias{kendallsTau,normalCopula-method}
\alias{kendallsTau,tCopula-method}
\alias{kendallsTau,claytonCopula-method}
\alias{kendallsTau,frankCopula-method}
\alias{kendallsTau,gumbelCopula-method}
\alias{spearmansRho}
\alias{spearmansRho-methods}
\alias{spearmansRho,normalCopula-method}
\alias{spearmansRho,tCopula-method}
\alias{spearmansRho,claytonCopula-method}
\alias{spearmansRho,frankCopula-method}
\alias{spearmansRho,gumbelCopula-method}
\title{Association Measures in Package `copula'}
\description{
 Methods to compute association measures such as Kendall's Tau and
 Spearman's Rho for bivaraite copulas.
}
\section{Methods}{
\describe{
\item{copula = "copula"}{ Association measure for a \code{"copula"} object.}
}}
\keyword{methods}
\keyword{multivariate}
