\name{copula-internal}
% Undocumented code objects

% Undocumented S4 classes:
\alias{schlatherCopula-class}
% Undocumented S4 methods
\title{Internal Copula Functions}
\description{
  Internal Copula functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
%\author{Marius Hofert, Ivan Kojadinovic, Martin Maechler and Jun Yan}
\keyword{internal}
