\name{fitCopula-class}
\docType{class}
\alias{fitCopula-class}
\alias{fitMvdc-class}
\alias{summaryFitCopula-class}
\alias{summaryFitMvdc-class}
\title{Class "fitCopula" }
\description{ Classes and summaries for fitting copula models. }
\section{Objects from the Class}{
  Objects can be created by calls of the form \code{fitCopula, fitMvdc}
  or their \code{summary} methods.
}
\author{ Jun Yan <jyan@stat.uiowa.edu> }
\keyword{classes}
