% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loglik.R
\name{get_loglik}
\alias{get_loglik}
\title{estimation of nested A. copula parameters}
\usage{
get_loglik(nac_Node, U, par)
}
\arguments{
\item{nac_Node}{an NAC node}

\item{U}{pseudo-observations}

\item{par}{a vector of parameters}
}
\value{
result obtained from optimizing provided parameters
}
\description{
estimation of nested A. copula parameters
}
