% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikNC.R
\name{loglikNC}
\alias{loglikNC}
\title{Estimation of nested A. copula parameters}
\usage{
loglikNC(par, U, naclist)
}
\arguments{
\item{par}{vector of new parameters}

\item{U}{matrix of pseudo-observations}

\item{naclist}{list specifying nesting structure}
}
\value{
sum of cdf of nested copula with new parameters
}
\description{
Estimation of nested A. copula parameters
}
