% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncopula-package.R
\docType{package}
\name{ncopula}
\alias{ncopula}
\alias{ncopula-package}
\title{ncopula - Nested copulas}
\description{
Estimation for nested copulas.
}
\author{
Amelia Tran <tran26h@mtholyoke.edu>, Evan Ray <eray@mtholyoke.edu>
}
