\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,copula-method}
\alias{show,normalCopula-method}
\alias{show,tCopula-method}
\title{  Methods for Function show in Package `copula' }
\description{
  Methods for function \code{show} in package \pkg{copula}.
}
\section{Methods}{
\describe{
\item{object = "copula"}{ ~~describe this method here }

\item{object = "normalCopula"}{ ~~describe this method here }

\item{object = "tCopula"}{ ~~describe this method here }
}}
\examples{
persp(tCopula(0.5, df = 25), dcopula)
persp(normalCopula(0.5), dcopula)
persp(tCopula(0.5, df = 25), pcopula)
persp(normalCopula(0.5), pcopula)
}
\keyword{methods}
\keyword{print}
