\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,normalCopula-method}
\alias{show,tCopula-method}
\alias{show,copula-method}
\alias{show,fitCopula-method}
\alias{show,fitMvdc-method}
\title{  Methods for function `show' in package `copula' }
\description{
  Methods for function \code{show} in package \pkg{copula}.
}
\section{Methods}{
  \describe{
    \item{object = "copula"}{see \code{\link{Copula}}. }
    
    \item{object = "fitCopula"}{see \code{\link{fitCopula}}. }
    
    \item{object = "fitMvdc"}{see \code{\link{fitCopula}}. }
}}
%\examples{

%}
\keyword{methods}
\keyword{print}
