\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,fitCopula-method}
\alias{summary,fitMvdc-method}
\title{  Methods for function `summary' in package `copula' }
\description{
  Methods for function \code{summary} in package \pkg{copula}.
}
\section{Methods}{
  \describe{
    \item{object = "fitCopula"}{see \code{\link{fitCopula}}.}
    
    \item{object = "fitMvdc"}{see \code{\link{fitCopula}}.}
}}
%\examples{

%}
\keyword{methods}

