\name{value-methods}
\docType{methods}
\alias{value-methods}
\alias{value,nACopula-method}
\title{Nested A.Copula Evaluation -- value() Methods}
\description{
  \code{value(<NAC>, u)} evaluates a nested Archimedean copula (an
  object of class \code{\linkS4class{nACopula}}) at \code{u}.
}
\section{Methods}{
\describe{
  \item{\code{signature(x = "nACopula")}}{ ~~describe this method here }
}}
\examples{
%% TODO
}
\keyword{methods}

